# sso/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/users/auth": {
            "post": {
                "description": "Validates identifier/password and returns JWT access token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authorization by login/email (get JWT)",
                "parameters": [
                    {
                        "description": "Credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.LoginResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/auth/phone": {
            "post": {
                "description": "Validates phone/password and returns JWT access token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Authorization by phone (get JWT)",
                "parameters": [
                    {
                        "description": "Phone credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.LoginByPhoneReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.LoginResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/google/callback": {
            "get": {
                "description": "Handles Google OAuth callback and returns JWT access token.\nIf FRONTEND_URL is set, redirects to FRONTEND_URL/auth/callback?token=\u003cJWT\u003e.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Google OAuth callback (get JWT)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Google auth code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON response when FRONTEND_URL is empty",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.LoginResp"
                        }
                    },
                    "302": {
                        "description": "Redirect to frontend callback with token when FRONTEND_URL is set",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/google/login": {
            "get": {
                "description": "Redirects user to Google OAuth consent screen",
                "tags": [
                    "users"
                ],
                "summary": "Google Login Redirect",
                "responses": {
                    "307": {
                        "description": "Redirect to Google",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Protected endpoint. Returns user data extracted from the Bearer token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "JWT verification / user info",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.AuthResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/otp/phone/confirm": {
            "post": {
                "description": "Confirms OTP and returns JWT access token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Confirm OTP by phone (get JWT)",
                "parameters": [
                    {
                        "description": "OTP confirm",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.ConfirmOTPByPhoneReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.LoginResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/otp/phone/send": {
            "post": {
                "description": "Sends OTP code to phone number.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Send OTP to phone",
                "parameters": [
                    {
                        "description": "Phone",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.SendOTPByPhoneReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/register": {
            "post": {
                "description": "Creates a user. Default role: user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.CreateUserReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.CreateUserResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/reset/email": {
            "post": {
                "description": "Sends OTP to email for password reset.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Reset password by email (send OTP)",
                "parameters": [
                    {
                        "description": "Email",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.ResetEmailReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/reset/email/confirm": {
            "post": {
                "description": "Confirms OTP and sets new password, returns JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Confirm email reset (OTP + new password)",
                "parameters": [
                    {
                        "description": "Confirm reset",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.ResetEmailConfirmReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.LoginResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/reset/phone": {
            "post": {
                "description": "Sends OTP to phone for password reset.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Reset password by phone (send OTP)",
                "parameters": [
                    {
                        "description": "Phone",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.ResetPhoneReq"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/users/reset/phone/confirm": {
            "post": {
                "description": "Confirms OTP and sets new password, returns JWT.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Confirm phone reset (OTP + new password)",
                "parameters": [
                    {
                        "description": "Confirm reset",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.ResetPhoneConfirmReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_handler.LoginResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_transport_http_handler.AuthResp": {
            "type": "object",
            "properties": {
                "role": {
                    "type": "string",
                    "example": "user"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_transport_http_handler.ConfirmOTPByPhoneReq": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1234
                },
                "phone": {
                    "type": "string",
                    "example": "+380501234567"
                }
            }
        },
        "internal_transport_http_handler.CreateUserReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "email@example.com"
                },
                "last_name": {
                    "type": "string",
                    "example": "lastName"
                },
                "login": {
                    "type": "string",
                    "example": "login"
                },
                "name": {
                    "type": "string",
                    "example": "name"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        },
        "internal_transport_http_handler.CreateUserResp": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_transport_http_handler.LoginByPhoneReq": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "phone": {
                    "type": "string",
                    "example": "+380501234567"
                }
            }
        },
        "internal_transport_http_handler.LoginReq": {
            "type": "object",
            "properties": {
                "identifier": {
                    "type": "string",
                    "example": "email@example.com"
                },
                "is_email": {
                    "type": "boolean",
                    "example": true
                },
                "password": {
                    "type": "string",
                    "example": "password"
                }
            }
        },
        "internal_transport_http_handler.LoginResp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "login": {
                    "type": "string",
                    "example": "login"
                },
                "role": {
                    "type": "string",
                    "example": "user"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "internal_transport_http_handler.ResetEmailConfirmReq": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1234
                },
                "email": {
                    "type": "string",
                    "example": "email@example.com"
                },
                "new_password": {
                    "type": "string",
                    "example": "newStrongPassword"
                }
            }
        },
        "internal_transport_http_handler.ResetEmailReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "email@example.com"
                }
            }
        },
        "internal_transport_http_handler.ResetPhoneConfirmReq": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1234
                },
                "new_password": {
                    "type": "string",
                    "example": "newStrongPassword"
                },
                "phone": {
                    "type": "string",
                    "example": "+380501234567"
                }
            }
        },
        "internal_transport_http_handler.ResetPhoneReq": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "+380501234567"
                }
            }
        },
        "internal_transport_http_handler.SendOTPByPhoneReq": {
            "type": "object",
            "properties": {
                "phone": {
                    "type": "string",
                    "example": "+380501234567"
                }
            }
        }
    }
}